+++ challenge_meli/setup.py
@@ -12,6 +12,7 @@ setup(
     name="challenge_meli",
     version=__version__,
     author="Luis Fillype",
+    python_requires=">=3.8.5",
     description="Data Engineering Challenge: MELI",
     long_description=readme,
     url="https://github.com/castielisgone/challenge-meli",
